backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "loguru>=0.7.3",
    "pydantic[email]>=2.11.7",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "supabase>=2.18.1",
    "uvicorn>=0.35.0",
]

[project.scripts]
start = "app.main:main"